refactor: update position of event functions and vars
